Cull some globals, use LooseVersion for all versions.
Globals are generally best avoided, and the call to platform is really
not going to be a performance hog, so I just call `get_os_version`
wherever the OS version is needed. It now returns a LooseVersion, to cut
down on the casting.

This also simnplifies one of the tests; before it was first breaking the
OS version into the major part, and then checking to see if it was
either 10.13 or 10.14. Now it just does >= 10.13.