Refactor main (#4)
* Move option parser creation and parsing to its own function.

* Cull some globals, use LooseVersion for all versions.

Globals are generally best avoided, and the call to platform is really
not going to be a performance hog, so I just call `get_os_version`
wherever the OS version is needed. It now returns a LooseVersion, to cut
down on the casting.

This also simnplifies one of the tests; before it was first breaking the
OS version into the major part, and then checking to see if it was
either 10.13 or 10.14. Now it just does >= 10.13.

* Alphabetize factored out optparser per Erik's systems.

* Compress simple function.

* Use tuple unpacking.

* Simplify conditional by testing membership.

* Factor duplicate NotificationCenter killer into func.

* Reduce repeeated string literal DRY offenses.

* Move duplicated timer call to after conditionals. Make it fit.

Used a common django idiom to break the chained method calls into
individual lines. When you're chaining a bunch together like this it
makes it a lot more readable to do each on an individual line.

* Clarify conditional, rewrap a line.

* Beautify another timer.

* Rename func and add docstring.