Skip notification db entries without dates
I discovered that Things.app creates entries in the com.apple.notificationcenter database with a Null date field, which causes a TypeError which drives you to Manual Enrollment even after successfully prompting a DEP nag. Don't know if other apps do this, but it seems safe to ignore entries without a date.